refactor: break out testify specifics
